requirements.txt
altgraph==0.17.4
antiorm==1.2.1
babel==2.17.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
db==0.1.1
db-sqlite3==0.0.1
dnspython==2.8.0
et_xmlfile==2.0.0
Flask==3.0.3
fonttools==4.57.0
greenlet==3.2.1
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
jdk4py==21.0.8.0
Jinja2==3.1.6
joblib==1.4.2
jsonfy==0.4
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
networkx==3.5
nltk==3.9.1
numpy==2.2.5
openapi-generator-cli==7.16.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pymongo==4.15.3
PyMySQL==1.1.2
pyodb==0.1.5
pyodbc==5.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
rake-nltk==1.0.6
regex==2024.11.6
reportlab==4.4.4
requests==2.32.5
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
spicy==0.16.0
SQLAlchemy==2.0.40
swaggerpy==0.2.1
threadpoolctl==3.6.0
tkcalendar==1.6.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websocket-client==1.9.0
Werkzeug==3.1.3
